BaseFragment 적용
